[CAP-352] - Miscellaneous task (#289)
* [CAP-532] - allow delete only draft job fair

* [CAP-532] - allow delete only draft job fair

* [CAP-532] - fix matching point service

+++ src/main/java/org/capstone/job_fair/services/impl/job_fair/JobFairServiceImpl.java
@@ -96,6 +96,8 @@ public class JobFairServiceImpl implements JobFairService {
             throw new IllegalArgumentException(MessageConstant.JobFair.JOB_FAIR_NOT_FOUND);
         }
         JobFairEntity entity = opt.get();
+        if (entity.getStatus() == JobFairPlanStatus.PUBLISH)
+            throw new IllegalArgumentException(MessageConstant.JobFair.JOB_FAIR_ALREADY_PUBLISH);
         jobFairRepository.delete(entity);
         return jobFairMapper.toDTO(entity);
     }
